src/requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
argon2-cffi
PyMySQL